requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
aw-client==0.5.9
aw-core==0.5.10
beautifulsoup4==4.11.1
bidict==0.22.0
browser-history==0.3.2
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
crashtest==0.3.1
customtkinter==5.0.1
darkdetect==0.7.1
dearpygui==1.8.0
deprecation==2.1.0
distlib==0.3.6
dnspython==2.2.1
dulwich==0.20.50
eventlet==0.33.2
exceptiongroup==1.0.4
filelock==3.8.0
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
greenlet==2.0.1
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==4.13.0
installer==0.5.1
iso8601==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
keyring==23.11.0
lockfile==0.12.2
MarkupSafe==2.1.2
metadata-parser==0.10.5
mock==4.0.3
more-itertools==9.0.0
MouseInfo==0.1.3
msgpack==1.0.4
multidict==6.0.3
numpy==1.24.1
outcome==1.2.0
packaging==21.3
pandas==1.5.3
peewee==3.15.4
pefile==2022.5.30
persist-queue==0.8.0
pexpect==4.8.0
Pillow==9.3.0
pkginfo==1.8.3
platformdirs==2.5.4
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pylev==1.4.0
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
PySocks==1.7.1
pystray==0.19.4
python-dateutil==2.8.2
python-engineio==4.3.4
python-json-logger==2.0.4
python-socketio==5.7.2
pytweening==1.0.4
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
scikit-learn==1.2.1
scipy==1.10.0
selenium==4.7.2
shellingham==1.5.0
six==1.16.0
sklearn==0.0.post2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
strict-rfc3339==0.7
tabulate==0.9.0
TakeTheTime==0.3.1
tendo==0.3.0
threadpoolctl==3.1.0
timeslot==0.1.2
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
uiautomation==2.0.17
urllib3==1.26.12
virtualenv==20.16.7
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
WMI==1.5.1
wsproto==1.2.0
yarl==1.8.2
zipp==3.10.0
